increment1 move all agent logic to one folder
